Add help messages to the build script
This might be just me, but figuring out how to run this script has cost
me 10 minutes of my time and some light lua debugging. This script has
a pretty poor user experience. In order to improve it I've added the
help message that is displayed whenever the build script is invoked
with at least one of the commonly-used options for getting help:

    -h, -help, or --help

Whenever the build script is executed without either -x64 or -a64
option I've also improved on the output of the script, telling what
exactly needs to be typed, and also directing the user to the real help
message. Since running the build script without any options and seeing
what happens is the first thing anyone looking at the build script will
do I see this as a good way of directing users to some kind of
documentation as soon as possible.

Minor thing, but I've also fixed the error code of the script when
running without the non-optional argument for arch. I think it's more
fitting to use the error code 2, which is commonly used for CLI-related
errors.

Signed-off-by: Sunagatov Denis <thebumboni@gmail.com>