Update CI actions' versions
The CI build failure was caused by the deprecation of v3 artifact
actions, which includes download-artifact@v3 and checkout@v3. See
https://github.com/orgs/community/discussions/142581 for details.

This fixes the versions of the GitHub actions so that the CI can pass.
The seanmiddleditch/gha-setup-ninja action has been explicitly set to
use @v6 version, since it's better for the thing to work than to break
randomly. Well it's an archived repo, so like what's the difference
anyway.

Signed-off-by: Sunagatov Denis <thebumboni@gmail.com>