ci: Add -x64 and -a64 to build.lua flags in jobs
Since the build.lua script now requires an architecture as the input we
need to be providing that in the CI script, which isn't. This builds
the TB and Cuik compilers with both x64 and a64 support (because for
now it is simpler than making a separate build per-architecture).

This also fixes the CI build script, as earlier the artifacts would not
be produced without architecture flag and the script would finish with
warnings and no artifacts.

Signed-off-by: Sunagatov Denis <thebumboni@gmail.com>